+++ packages/apidom-converter/test/strategies/openapi-3-1-to-openapi-3-0-3/refractor-plugins/security-requirements-empty-roles/index.ts
@@ -2,7 +2,7 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { mediaTypes as openAPI31MediaTypes } from '@swagger-api/apidom-parser-adapter-openapi-json-3-1';
 import { mediaTypes as openAPI30MediaTypes } from '@swagger-api/apidom-parser-adapter-openapi-json-3-0';
-import { AnnotationElement, includesClasses, toJSON, toValue } from '@swagger-api/apidom-core';
+import { AnnotationElement, includesClasses, toJSON } from '@swagger-api/apidom-core';
 import { assert, expect } from 'chai';
 
 import convert from '../../../../../src/index.ts';
@@ -73,20 +73,22 @@ describe('converter', function () {
           const annotation: AnnotationElement = annotations.find((a: AnnotationElement) =>
             a.code?.equals('security-requirements-empty-roles'),
           );
-          const sourceMap = annotation.meta.get('sourceMap');
-          const { positionStart, positionEnd } = sourceMap;
-          const [startRow, startColumn, startChar] = toValue(positionStart);
-          const [endRow, endColumn, endChar] = toValue(positionEnd);
+          const {
+            startPositionRow,
+            startPositionColumn,
+            startIndex,
+            endPositionRow,
+            endPositionColumn,
+            endIndex,
+          } = annotation;
 
-          assert.isDefined(sourceMap);
+          assert.strictEqual(startPositionRow, 15);
+          assert.strictEqual(startPositionColumn, 29);
+          assert.strictEqual(startIndex, 299);
 
-          assert.strictEqual(startRow, 15);
-          assert.strictEqual(startColumn, 29);
-          assert.strictEqual(startChar, 299);
-
-          assert.strictEqual(endRow, 18);
-          assert.strictEqual(endColumn, 13);
-          assert.strictEqual(endChar, 358);
+          assert.strictEqual(endPositionRow, 18);
+          assert.strictEqual(endPositionColumn, 13);
+          assert.strictEqual(endIndex, 358);
         });
       });
     });