+++ packages/apidom-parser-adapter-json/src/syntactic-analysis/indirect/visitors/CstVisitor.ts
@@ -11,11 +11,11 @@ import {
   JsonStringContent,
   JsonTrue,
   ParseResult,
-  Position,
-  Point,
   Literal,
   Error,
 } from '@swagger-api/apidom-ast';
+import type { JsonValueOptions } from '@swagger-api/apidom-ast';
+import { assignSourceMap } from '@swagger-api/apidom-core';
 
 import TreeCursorSyntaxNode from '../../TreeCursorSyntaxNode.ts';
 
@@ -32,30 +32,11 @@ export const keyMap = {
 /* eslint-disable class-methods-use-this */
 
 class CstVisitor {
-  private static toPosition(node: TreeCursorSyntaxNode): Position {
-    const start = new Point({
-      row: node.startPosition.row,
-      column: node.startPosition.column,
-      char: node.startIndex,
-    });
-    const end = new Point({
-      row: node.endPosition.row,
-      column: node.endPosition.column,
-      char: node.endIndex,
-    });
-
-    return new Position({ start, end });
-  }
-
   public readonly document = {
     enter: (node: TreeCursorSyntaxNode): JsonDocument => {
-      const position = CstVisitor.toPosition(node);
-
-      return new JsonDocument({
-        children: node.children,
-        position,
-        isMissing: node.isMissing,
-      });
+      return new JsonDocument(
+        assignSourceMap({ children: node.children, isMissing: node.isMissing }, node),
+      );
     },
     leave: (document: JsonDocument): ParseResult => {
       return new ParseResult({ children: [document] });
@@ -65,77 +46,91 @@ class CstVisitor {
   public enter(node: TreeCursorSyntaxNode): Literal | undefined {
     // anonymous literals from CST transformed into AST literal nodes
     if (node instanceof TreeCursorSyntaxNode && !node.isNamed) {
-      const position = CstVisitor.toPosition(node);
       const value = node.type || node.text;
       const { isMissing } = node;
 
-      return new Literal({ value, position, isMissing });
+      return new Literal(assignSourceMap({ value, isMissing }, node));
     }
 
     return undefined;
   }
 
   public object(node: TreeCursorSyntaxNode): JsonObject {
-    const position = CstVisitor.toPosition(node);
-
-    return new JsonObject({ children: node.children, position, isMissing: node.isMissing });
+    return new JsonObject(
+      assignSourceMap({ children: node.children, isMissing: node.isMissing }, node),
+    );
   }
 
   public pair(node: TreeCursorSyntaxNode): JsonProperty {
-    const position = CstVisitor.toPosition(node);
     const children = node.children.slice(1);
     const { keyNode } = node;
-    const key = new JsonKey({
-      children: keyNode?.children || [],
-      position: keyNode != null ? CstVisitor.toPosition(keyNode) : undefined,
-      isMissing: keyNode != null ? keyNode.isMissing : false,
-    });
+    const key = new JsonKey(
+      assignSourceMap(
+        {
+          children: keyNode?.children || [],
+          isMissing: keyNode != null ? keyNode.isMissing : false,
+        },
+        keyNode,
+      ),
+    );
 
-    return new JsonProperty({ children: [key, ...children], position, isMissing: node.isMissing });
+    return new JsonProperty(
+      assignSourceMap(
+        {
+          children: [key, ...children],
+          isMissing: node.isMissing,
+        },
+        node,
+      ),
+    );
   }
 
   public array(node: TreeCursorSyntaxNode): JsonArray {
-    const position = CstVisitor.toPosition(node);
-
-    return new JsonArray({ children: node.children, position, isMissing: node.isMissing });
+    return new JsonArray(
+      assignSourceMap({ children: node.children, isMissing: node.isMissing }, node),
+    );
   }
 
   public string(node: TreeCursorSyntaxNode): JsonString {
-    const position = CstVisitor.toPosition(node);
     const content = new JsonStringContent({ value: JSON.parse(node.text) });
 
-    return new JsonString({ children: [content], position, isMissing: node.isMissing });
+    return new JsonString(
+      assignSourceMap({ children: [content], isMissing: node.isMissing }, node),
+    );
   }
 
   public number(node: TreeCursorSyntaxNode): JsonNumber {
-    const position = CstVisitor.toPosition(node);
     const value = node.text;
-
-    return new JsonNumber({ value, position, isMissing: node.isMissing });
+    return new JsonNumber(
+      assignSourceMap({ value, isMissing: node.isMissing }, node) as JsonValueOptions,
+    );
   }
 
   // eslint-disable-next-line @typescript-eslint/naming-convention
   public null(node: TreeCursorSyntaxNode): JsonNull {
-    const position = CstVisitor.toPosition(node);
     const value = node.text;
 
-    return new JsonNull({ value, position, isMissing: node.isMissing });
+    return new JsonNull(
+      assignSourceMap({ value, isMissing: node.isMissing }, node) as JsonValueOptions,
+    );
   }
 
   // eslint-disable-next-line @typescript-eslint/naming-convention
   public true(node: TreeCursorSyntaxNode): JsonTrue {
-    const position = CstVisitor.toPosition(node);
     const value = node.text;
 
-    return new JsonTrue({ value, position, isMissing: node.isMissing });
+    return new JsonTrue(
+      assignSourceMap({ value, isMissing: node.isMissing }, node) as JsonValueOptions,
+    );
   }
 
   // eslint-disable-next-line @typescript-eslint/naming-convention
   public false(node: TreeCursorSyntaxNode): JsonFalse {
-    const position = CstVisitor.toPosition(node);
     const value = node.text;
 
-    return new JsonFalse({ value, position, isMissing: node.isMissing });
+    return new JsonFalse(
+      assignSourceMap({ value, isMissing: node.isMissing }, node) as JsonValueOptions,
+    );
   }
 
   public ERROR(
@@ -144,14 +139,17 @@ class CstVisitor {
     parent: unknown,
     path: string[],
   ): ParseResult | Error {
-    const position = CstVisitor.toPosition(node);
-    const errorNode = new Error({
-      children: node.children,
-      position,
-      isUnexpected: !node.hasError,
-      isMissing: node.isMissing,
-      value: node.text,
-    });
+    const errorNode = new Error(
+      assignSourceMap(
+        {
+          children: node.children,
+          isUnexpected: !node.hasError,
+          isMissing: node.isMissing,
+          value: node.text,
+        },
+        node,
+      ),
+    );
 
     if (path.length === 0) {
       return new ParseResult({ children: [errorNode] });