+++ packages/apidom-parser-adapter-yaml-1-2/test/adapter-node/empty-nodes/mappings/block-maping/empty-value-properties.ts
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import { includesClasses, toValue, SourceMapElement } from '@swagger-api/apidom-core';
+import { includesClasses, hasElementSourceMap } from '@swagger-api/apidom-core';
 
 import * as adapter from '../../../../../src/adapter-node.ts';
 
@@ -19,36 +19,25 @@ describe('given empty node with tag and anchor as block mapping value', function
 
   it('should generate source maps', async function () {
     const emptyElement = await setupEmptyElement();
-    const sourceMapElement = emptyElement.meta.get('sourceMap');
 
-    assert.instanceOf(sourceMapElement, SourceMapElement);
+    assert.isTrue(hasElementSourceMap(emptyElement));
   });
 
   it('should generate proper source map start position', async function () {
     const emptyElement = await setupEmptyElement();
-    const sourceMapElement = emptyElement.meta.get('sourceMap');
-    const [row, column, char] = [
-      toValue(sourceMapElement.positionStart.get(0)),
-      toValue(sourceMapElement.positionStart.get(1)),
-      toValue(sourceMapElement.positionStart.get(2)),
-    ];
+    const { startPositionRow, startPositionColumn, startIndex } = emptyElement;
 
-    assert.strictEqual(row, 0);
-    assert.strictEqual(column, 17);
-    assert.strictEqual(char, 17);
+    assert.strictEqual(startPositionRow, 0);
+    assert.strictEqual(startPositionColumn, 17);
+    assert.strictEqual(startIndex, 17);
   });
 
   it('should generate proper source map end position', async function () {
     const emptyElement = await setupEmptyElement();
-    const sourceMapElement = emptyElement.meta.get('sourceMap');
-    const [row, column, char] = [
-      toValue(sourceMapElement.positionEnd.get(0)),
-      toValue(sourceMapElement.positionEnd.get(1)),
-      toValue(sourceMapElement.positionEnd.get(2)),
-    ];
+    const { endPositionRow, endPositionColumn, endIndex } = emptyElement;
 
-    assert.strictEqual(row, 0);
-    assert.strictEqual(column, 17);
-    assert.strictEqual(char, 17);
+    assert.strictEqual(endPositionRow, 0);
+    assert.strictEqual(endPositionColumn, 17);
+    assert.strictEqual(endIndex, 17);
   });
 });